pyproject.toml
[project]
name = "astrostudy"
version = "0.1.0"
description = "Asteroid risk analysis using NASA NeoWs API"
requires-python = ">=3.11"
authors = [
    {name = "AstroStudy Team"}
]

dependencies = [
    "requests>=2.31",
    "pydantic>=2.0",
    "pandas>=2.0",
    "scikit-learn>=1.4",
    "joblib>=1.3",
    "streamlit>=1.35",
    "plotly>=5.20",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "loguru>=0.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "ipykernel>=6.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "UP"]

[tool.mypy]
ignore_missing_imports = true
